+++ app/presenters/report_presenter.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-class SubmissionPresenter < SimpleDelegator
+class ReportPresenter < SimpleDelegator
   def state
     return if super.blank?
-    I18n.t("label.submission.state.#{super.downcase}")
+    I18n.t("label.report.state.#{super.downcase}")
   end
 
   def deadline